Added to the cake build and cleaned up unneeded classes
