+++ src/NUnitEngine/Portable/nunit.driver.tests/nunit.driver.tests.csproj
@@ -39,8 +39,9 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="NUnit3PortableDriverTests.cs" />
+    <Compile Include="NUnitDriverTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="XmlHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\NUnitFramework\framework\nunit.framework-4.5.csproj">
